+++ packages/viewer/web/src/utils/convertContentState.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { renderToStaticMarkup } from 'react-dom/server';
+
 import {
   BLOCK_TYPES,
   depthClassName,
@@ -252,6 +252,7 @@ const convertToReact = (
   );
 };
 
+// renderToStaticMarkup param should be imported 'react-dom/server' (in order reduce viewer bundle size and probably not used anyhow)
 const convertToHTML = (
   contentState,
   mergedStyles,
@@ -259,6 +260,7 @@ const convertToHTML = (
   typeMap,
   entityProps,
   decorators,
+  renderToStaticMarkup,
   options = {}
 ) => {
   if (isEmptyContentState(contentState)) {