+++ CHANGELOG.md
@@ -15,6 +15,10 @@
     Changes that have landed in master but are not yet released.
     Click to see more.
   </summary>
+
+### :bug: Bug Fix
+- `image`
+  - [#1399](https://github.com/wix-incubator/rich-content/pull/1399) link in image viewer
 </details>
 <hr/>