+++ CHANGELOG.md
@@ -21,6 +21,7 @@
   - [#1399](https://github.com/wix-incubator/rich-content/pull/1399) link in image viewer
 - `viewer`
   - [#1395](https://github.com/wix-incubator/rich-content/pull/1395) remove react-dom/server from viewer dependencies (reduce bundle size)
+  - [#1404](https://github.com/wix-incubator/rich-content/pull/1404) change viewer justify css
 </details>
 <hr/>